requirements.txt
requests==2.31.0
PyYAML==6.0.1
pyJWT==2.8.0
Pillow==10.1.0

black
isort
